+++ molgenis-app-lifelines/src/main/java/org/molgenis/lifelines/plugins/HeaderPlugin.java
@@ -1,4 +1,4 @@
-package org.molgenis.lifelines.plugins.header;
+package org.molgenis.lifelines.plugins;
 
 import java.io.Serializable;
 
@@ -12,7 +12,7 @@ import org.molgenis.util.ApplicationContextProvider;
 import org.molgenis.util.Entity;
 import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 
-public class Header extends PluginModel<Entity>
+public class HeaderPlugin extends PluginModel<Entity>
 {
 	private static final long serialVersionUID = 1L;
 
@@ -23,7 +23,7 @@ public class Header extends PluginModel<Entity>
 
 	private HeaderModel headerModel;
 
-	public Header(String name, ScreenController<?> parent)
+	public HeaderPlugin(String name, ScreenController<?> parent)
 	{
 		super(name, parent);
 	}
@@ -49,7 +49,7 @@ public class Header extends PluginModel<Entity>
 	@Override
 	public String getViewTemplate()
 	{
-		return "templates/" + Header.class.getName().replace('.', '/') + ".ftl";
+		return "templates/" + HeaderPlugin.class.getName().replace('.', '/') + ".ftl";
 	}
 
 	@Override
