+++ molgenis-app-lifelines/src/main/webapp/WEB-INF/classes/templates/org/molgenis/lifelines/plugins/HeaderPlugin.ftl
