+++ molgenis-app-omx/src/main/java/org/molgenis/omx/plugins/HomePlugin.java
@@ -1,4 +1,6 @@
-package org.molgenis.omx.plugins.home;
+package org.molgenis.omx.plugins;
+
+import java.io.Serializable;
 
 import org.molgenis.framework.db.Database;
 import org.molgenis.framework.server.MolgenisRequest;
@@ -9,7 +11,7 @@ import org.molgenis.util.ApplicationContextProvider;
 import org.molgenis.util.Entity;
 import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 
-public class Home extends PluginModel<Entity>
+public class HomePlugin extends PluginModel<Entity>
 {
 	private static final long serialVersionUID = 1L;
 
@@ -23,7 +25,7 @@ public class Home extends PluginModel<Entity>
 		return homeModel;
 	}
 
-	public Home(String name, ScreenController<?> parent)
+	public HomePlugin(String name, ScreenController<?> parent)
 	{
 		super(name, parent);
 	}
@@ -31,13 +33,13 @@ public class Home extends PluginModel<Entity>
 	@Override
 	public String getViewName()
 	{
-		return Home.class.getSimpleName();
+		return HomePlugin.class.getSimpleName();
 	}
 
 	@Override
 	public String getViewTemplate()
 	{
-		return "templates/" + Home.class.getName().replace('.', '/') + ".ftl";
+		return "templates/" + HomePlugin.class.getName().replace('.', '/') + ".ftl";
 	}
 
 	@Override
@@ -74,8 +76,10 @@ public class Home extends PluginModel<Entity>
 		}
 	}
 
-	public static class HomeModel
+	public static class HomeModel implements Serializable
 	{
+		private static final long serialVersionUID = 1L;
+
 		private String homeHtml;
 
 		public String getHomeHtml()
