+++ molgenis-app-omx/src/main/webapp/WEB-INF/classes/templates/org/molgenis/omx/plugins/HeaderPlugin.ftl
